refactor: Break-up top-level app (single responsibility)
